dom/imptests/html/dom/nodes/test_Document-createComment.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Document.createComment</title>
<link rel=help href="http://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html#dom-document-createcomment">
<link rel=help href="http://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html#dom-node-ownerdocument">
<link rel=help href="http://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html#dom-characterdata-data">
<link rel=help href="http://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html#dom-node-nodevalue">
<link rel=help href="http://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html#dom-node-textcontent">
<link rel=help href="http://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html#dom-characterdata-length">
<link rel=help href="http://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html#dom-node-nodetype">
<link rel=help href="http://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html#dom-node-haschildnodes">
<link rel=help href="http://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html#dom-node-childnodes">
<link rel=help href="http://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html#dom-node-firstchild">
<link rel=help href="http://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html#dom-node-lastchild">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  ["\u000b", "a -- b", "a-", "-b"].forEach(function(value) {
    test(function() {
      var c = document.createComment(value);
      assert_true(c instanceof Comment);
      assert_true(c instanceof Node);
      assert_equals(c.ownerDocument, document);
      assert_equals(c.data, value);
      assert_equals(c.nodeValue, value);
      assert_equals(c.textContent, value);
      assert_equals(c.length, value.length);
      assert_equals(c.nodeType, 8);
      assert_equals(c.nodeName, "#comment");
      assert_equals(c.hasChildNodes(), false);
      assert_equals(c.childNodes.length, 0);
      assert_equals(c.firstChild, null);
      assert_equals(c.lastChild, null);
    }, "createComment(" + format_value(value) + ")");
  });
});
</script>
